+++ src/test/java/org/opentripplanner/netex/mapping/RouteMapperTest.java
@@ -29,10 +29,10 @@ import org.rutebanken.netex.model.GroupOfLinesRefStructure;
 import org.rutebanken.netex.model.Line;
 import org.rutebanken.netex.model.MultilingualString;
 import org.rutebanken.netex.model.Network;
-import org.rutebanken.netex.model.OrganisationRefStructure;
 import org.rutebanken.netex.model.PresentationStructure;
+import org.rutebanken.netex.model.TransportOrganisationRefStructure;
 
-public class RouteMapperTest {
+class RouteMapperTest {
 
   private static final String NETWORK_ID = "RUT:Network:1";
   private static final String GOL_ID_1 = "RUT:GroupOfLines:1";
@@ -47,7 +47,7 @@ public class RouteMapperTest {
   private static final Set<String> EMPTY_FERRY_WITHOUT_BICYCLE_IDS = Collections.emptySet();
 
   @Test
-  public void mapRouteWithDefaultAgency() {
+  void mapRouteWithDefaultAgency() {
     NetexEntityIndex netexEntityIndex = new NetexEntityIndex();
     Line line = createExampleLine();
 
@@ -72,7 +72,7 @@ public class RouteMapperTest {
   }
 
   @Test
-  public void mapRouteWithAgencySpecified() {
+  void mapRouteWithAgencySpecified() {
     NetexEntityIndex netexIndex = new NetexEntityIndex();
     OtpTransitServiceBuilder transitBuilder = new OtpTransitServiceBuilder(
       DataImportIssueStore.NOOP
@@ -81,7 +81,7 @@ public class RouteMapperTest {
     Network network = new Network()
       .withId(NETWORK_ID)
       .withTransportOrganisationRef(
-        createJaxbElement(new OrganisationRefStructure().withRef(AUTHORITY_ID))
+        createJaxbElement(new TransportOrganisationRefStructure().withRef(AUTHORITY_ID))
       );
 
     netexIndex.networkById.add(network);
@@ -110,7 +110,7 @@ public class RouteMapperTest {
   }
 
   @Test
-  public void mapRouteWithColor() {
+  void mapRouteWithColor() {
     NetexEntityIndex netexEntityIndex = new NetexEntityIndex();
     Line line = createExampleLine();
     byte[] color = new byte[] { 127, 0, 0 };
@@ -132,12 +132,12 @@ public class RouteMapperTest {
 
     Route route = routeMapper.mapRoute(line);
 
-    assertEquals(route.getColor(), "7F0000");
-    assertEquals(route.getTextColor(), "007F00");
+    assertEquals("7F0000", route.getColor());
+    assertEquals("007F00", route.getTextColor());
   }
 
   @Test
-  public void allowBicyclesOnFerries() {
+  void allowBicyclesOnFerries() {
     NetexEntityIndex netexEntityIndex = new NetexEntityIndex();
     Line lineWithBicycles = createExampleFerry(LINE_ID);
     Line lineWithOutBicycles = createExampleFerry(FERRY_WITHOUT_BICYCLES_ID);
@@ -163,7 +163,7 @@ public class RouteMapperTest {
   }
 
   @Test
-  public void mapRouteWithoutBranding() {
+  void mapRouteWithoutBranding() {
     NetexEntityIndex netexEntityIndex = new NetexEntityIndex();
     Line line = createExampleLine();
 
@@ -186,7 +186,7 @@ public class RouteMapperTest {
   }
 
   @Test
-  public void mapRouteWithBranding() {
+  void mapRouteWithBranding() {
     NetexEntityIndex netexIndex = new NetexEntityIndex();
     OtpTransitServiceBuilder transitBuilder = new OtpTransitServiceBuilder(
       DataImportIssueStore.NOOP
@@ -219,7 +219,7 @@ public class RouteMapperTest {
   }
 
   @Test
-  public void mapRouteWithGroupOfRoutes() {
+  void mapRouteWithGroupOfRoutes() {
     NetexEntityIndex netexIndex = new NetexEntityIndex();
     OtpTransitServiceBuilder transitBuilder = new OtpTransitServiceBuilder(
       DataImportIssueStore.NOOP